Polished Spring AOP advice types question, corrected integration tests
