+++ src/main/java/spring/certification/aop/q005/example/AdviceConfigurationMarker.java
@@ -1,0 +1,11 @@
+package spring.certification.aop.q005.example;
+
+import org.springframework.context.annotation.ComponentScan;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.context.annotation.EnableAspectJAutoProxy;
+
+@Configuration
+@ComponentScan
+@EnableAspectJAutoProxy
+public class AdviceConfigurationMarker {
+}
