+++ src/main/java/spring/certification/aop/q005/example/PanicException.java
@@ -1,0 +1,14 @@
+package spring.certification.aop.q005.example;
+
+/**
+ * Sometimes people panic.
+ */
+public class PanicException extends RuntimeException {
+
+    public PanicException() {
+    }
+
+    public PanicException(String message) {
+        super(message);
+    }
+}
