+++ src/main/java/spring/certification/aop/q005/example/annotation/GiveMeAfterCue.java
@@ -1,0 +1,11 @@
+package spring.certification.aop.q005.example.annotation;
+
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
+
+@Target(ElementType.METHOD)
+@Retention(RetentionPolicy.RUNTIME)
+public @interface GiveMeAfterCue {
+}
