+++ src/main/java/spring/certification/aop/q005/example/annotation/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * Marker annotations for methods which ask to 'advice' them.
+ */
+package spring.certification.aop.q005.example.annotation;
